Rewrite argument parsing to use Clap's derives
This is something that has been on the radar for a while, but the recent
deprecation of everything from Clap 3.1 in Clap 3.2 has pushed our hand
a bit to get it done now with the new best practices from Clap 3.2.

+++ src/api/config/speedsettings.rs
@@ -112,7 +112,7 @@ impl Default for SpeedSettings {
 
 impl SpeedSettings {
   /// Set the speed setting according to a numeric speed preset.
-  pub fn from_preset(speed: usize) -> Self {
+  pub fn from_preset(speed: u8) -> Self {
     // The default settings are equivalent to speed 0
     let mut settings = SpeedSettings::default();
